+++ ParadeDB.EntityFrameworkCore.slnx
@@ -47,6 +47,6 @@
     <Project Path="src\ParadeDB.EntityFrameworkCore\ParadeDB.EntityFrameworkCore.csproj" />
   </Folder>
   <Folder Name="/tests/">
-    <Project Path="tests/ParadeDB.EntityFrameworkCore.IntegrationTests.csproj" />
+    <Project Path="tests/ParadeDB.EntityFrameworkCore.Tests.csproj" />
   </Folder>
 </Solution>
