+++ tests/ParadeDB.EntityFrameworkCore.Tests.csproj
